backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Shravan Asati <dev.shravan@protonmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
nasapy = "^0.2.7"
redis = {extras = ["hiredis"], version = "^5.0.8"}
pandas = "^2.2.2"
folium = "^0.17.0"
requests = "^2.32.3"
google-generativeai = "^0.7.2"
gunicorn = "^23.0.0"
flask-limiter = "^3.8.0"

[tool.poetry.group.dev]  # This part can be left out

[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
